Validacion de menues
se adiciono validaciones de los distintos menues, y se añadieron nuevas validades al cuit

+++ SistemaDeStock/Cliente.h
@@ -7,18 +7,18 @@ class Cliente{
 private:
     int _ID;
     char _NombreEmpresa[50];
-    int _Cuit;
+    long long _Cuit;
     char _Correo[50];
     bool _Estado;
 public:
     void setID(int id);
     void setNombreEmpresa(string NombreEmpresa);
-    void setCuit(int Cuit);
+    void setCuit(long long Cuit);
     void setCorreo(string Correo);
     void setEstado(bool Estado);
     int getID();
     string getNombreEmpresa();
-    int getCuit();
+    long long getCuit();
     string getCorreo();
     bool getEstado();
     void Cargar();
